Delay blur image until cover is loaded
